+++ lib/Parse/Scope.cpp
@@ -15,7 +15,6 @@
 //===----------------------------------------------------------------------===//
 
 #include "swift/Parse/Scope.h"
-#include "swift/AST/Attr.h"
 #include "swift/Parse/Parser.h"
 #include "llvm/ADT/Twine.h"
